requirements.txt
Flask==3.0.3
google-generativeai==0.3.0
pymongo==4.7.1
gunicorn==21.2.0 
dnspython
certifi